+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Unreleased
+
+#### Fixed
+
+* Remove invalid HTML from selected list input to avoid error in Rails 7.2 [#507](https://github.com/platanus/activeadmin_addons/pull/507)
+
 ### 1.10.1
 
 * Backport [#477](https://github.com/platanus/activeadmin_addons/pull/477) to have ActiveAdmin v3 compatibility [#479](https://github.com/platanus/activeadmin_addons/pull/479)